refactor: move o11y
