+++ src/infrastructures/observability/logging/client-logger.ts
